requirements.txt
anyio==4.12.1
APScheduler==3.11.2
av==16.1.0
certifi==2026.1.4
click==8.3.1
coloredlogs==15.0.1
ctranslate2==4.6.3
faster-whisper==1.2.1
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.5
humanfriendly==10.0
idna==3.11
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mpmath==1.3.0
numpy==2.4.2
onnxruntime==1.23.2
packaging==26.0
platformdirs==4.5.1
protobuf==6.33.5
Pygments==2.19.2
python-dotenv==1.2.1
python-telegram-bot==22.6
PyYAML==6.0.3
rich==14.3.2
setuptools==80.10.2
shellingham==1.5.4
sympy==1.14.0
textual==7.5.0
tokenizers==0.22.2
tqdm==4.67.2
typer-slim==0.21.1
typing_extensions==4.15.0
tzlocal==5.3.1
uc-micro-py==1.0.3
yt-dlp==2023.11.16
pyyaml
pymupdf
python-docx